% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{crear_paleta_colores}
\alias{crear_paleta_colores}
\title{Crea una paleta interpolada basada en combinación Wada Sanzo}
\usage{
crear_paleta_colores(id, n = 6)
}
\arguments{
\item{id}{Número o carácter con el ID de combinación}

\item{n}{Número de colores a generar (por defecto 6)}
}
\value{
Vector de colores HEX interpolados
}
\description{
Crea una paleta interpolada basada en combinación Wada Sanzo
}
