% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{.preparar_diccionario_combinaciones}
\alias{.preparar_diccionario_combinaciones}
\title{Prepara el diccionario de combinaciones a partir del JSON}
\usage{
.preparar_diccionario_combinaciones(json)
}
\arguments{
\item{json}{Lista con datos crudos del JSON}
}
\value{
Dataframe con combinaciones y sus colores hex
}
\description{
Prepara el diccionario de combinaciones a partir del JSON
}
