% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{obtener_combinacion_wada}
\alias{obtener_combinacion_wada}
\title{Obtiene los colores y nombres de una combinación Wada Sanzo}
\usage{
obtener_combinacion_wada(id_combinacion)
}
\arguments{
\item{id_combinacion}{Número o carácter con el ID de combinación}
}
\value{
Lista con elementos `colores` (vector HEX) y `nombres` (vector de nombres)
}
\description{
Obtiene los colores y nombres de una combinación Wada Sanzo
}
